upgrade mermaid
