+++ package.json
@@ -85,7 +85,7 @@
     "lodash": "^4.18.1",
     "markdown-to-txt": "^2.0.1",
     "markdown-toc": "^1.2.0",
-    "mermaid": "9.3.0",
+    "mermaid": "^11.4.1",
     "next": "15.5.21",
     "next-i18next": "^15.3.0",
     "next-language-detector": "^1.1.1",
